FINAL/requirements.txt
streamlit
requests
python-dotenv
langserve
fastapi
uvicorn
sentence-transformers
langchain-pinecone 
pinecone-notebooks
jq
chromadb
langchain_community
openai==0.28